region-captioning-main/region-captioning-main/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy
nltk
Pillow
tqdm
pandas
ultralytics
pycocotools
pycocoevalcap